Modified Visualizer code and removed unused files.

+++ app/src/main/java/solutiontogo/de/audiocitytourguide/recording/AudioRecordingHandler.java
@@ -5,5 +5,5 @@ package solutiontogo.de.audiocitytourguide.recording;
  */
 
 public interface AudioRecordingHandler {
-    void updateVisualizerView(float amplitude);
+    void updateVisualizerView(int amplitude);
 }
